pyproject.toml
[tool.poetry]
name = "upload-all-images-in-markdown"
version = "0.1.0"
description = "convert local image links to online markdown link, and upload via picgo services."
authors = ["siben <siben168@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[tool.poetry.scripts]
upload-all-images-in-markdown = "upload_all_images_in_markdown:main"

[tool.poetry.group.dev.dependencies]
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
